requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
coloredlogs==15.0.1
datasets==2.12.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.6
docarray==0.21.0
docker==6.1.3
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.97.0
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.6.0
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
googleapis-common-protos==1.59.1
grpcio==1.47.5
grpcio-health-checking==1.47.5
grpcio-reflection==1.47.5
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.15.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.1
jcloud==0.2.11
jieba==0.42.1
jina==3.17.0
jina-hubble-sdk==0.38.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
lit==16.0.5.post0
loguru==0.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnxruntime==1.15.0
opencv-python==4.7.0.72
opentelemetry-api==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-aiohttp-client==0.39b0
opentelemetry-instrumentation-asgi==0.39b0
opentelemetry-instrumentation-fastapi==0.39b0
opentelemetry-instrumentation-grpc==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
piexif==1.1.3
Pillow==9.5.0
prometheus-client==0.17.0
protobuf==4.23.2
pyarrow==12.0.1
pyasn1==0.5.0
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
responses==0.18.0
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
streamlit==1.23.1
sympy==1.12
tenacity==8.2.2
text2vec==1.1.8
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.19.0
websocket-client==1.5.3
websockets==11.0.3
wrapt==1.15.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
